Fix all warnings
0 warnings :)

+++ src/utils/OpenBrowser.cpp
@@ -9,6 +9,6 @@ namespace EGL3::Utils {
     // Split into a seperate file so including this header doesn't include windows.h
     bool OpenInBrowser(const std::string& Url)
     {
-        return (int)ShellExecuteA(NULL, NULL, Url.c_str(), NULL, NULL, SW_SHOW) > 32;
+        return (uint64_t)ShellExecuteA(NULL, NULL, Url.c_str(), NULL, NULL, SW_SHOW) > 32;
     }
 }
